+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # SQL Build Manager Change Log
 
 ### Version 15.2.2-beta
+- *NEW:* Eliminated the need for `sbm batch enqueue`. You can now run `sbm batch run` and it will automatically enqueue the database targets for you. You can still run `sbm batch enqueue` first if desired
 - *ADDED:* EventHub logging now also includes the script results for each script run against the databases.  
 - *ADDED:* New option `--eventhublogging` controls how to log script results and if to emit verbose message events. Add multiple flags to combine settings. Values: EssentialOnly|IndividualScriptResults|ConsolidatedScriptResults|VerboseMessages")
 - *FIXED:* Corrected bug where Batch execution wasn't properly consolidating certain log files